Move Runtime Input tests to a separate assembly
