+++ Project/ProjectSettings/SceneTemplateSettings.json
@@ -63,6 +63,11 @@
         },
         {
             "userAdded": false,
+            "type": "UnityEngine.PhysicsMaterial",
+            "defaultInstantiationMode": 0
+        },
+        {
+            "userAdded": false,
             "type": "UnityEngine.PhysicsMaterial2D",
             "defaultInstantiationMode": 0
         },
