+++ com.unity.ml-agents/Tests/Runtime/Input/Adaptors/FloatInputActionAdapatorTests.cs
@@ -4,7 +4,7 @@ using Unity.MLAgents.Actuators;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace Unity.MLAgents.Tests.Actuators
+namespace Unity.MLAgents.Tests.Input
 {
     public class FloatInputActionAdaptorTests : InputTestFixture
     {
