+++ com.unity.ml-agents/Tests/Runtime/Input/Adaptors/Vector2InputActionAdaptorTests.cs
@@ -4,7 +4,7 @@ using Unity.MLAgents.Actuators;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace Unity.MLAgents.Tests.Actuators
+namespace Unity.MLAgents.Tests.Input
 {
     public class Vector2InputActionAdaptorTests : InputTestFixture
     {
