+++ com.unity.ml-agents/Tests/Runtime/Input/Unity.ML-Agents.Runtime.Input.Tests.asmdef
@@ -1,0 +1,37 @@
+{
+    "name": "Unity.ML-Agents.Runtime.Input.Tests",
+    "rootNamespace": "",
+    "references": [
+        "Unity.ML-Agents",
+        "Unity.ML-Agents.CommunicatorObjects",
+        "UnityEngine.TestRunner",
+        "UnityEditor.TestRunner",
+        "Unity.InferenceEngine",
+        "Unity.InputSystem",
+        "Unity.InputSystem.TestFramework"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "nunit.framework.dll"
+    ],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_INCLUDE_TESTS"
+    ],
+    "versionDefines": [
+        {
+            "name": "com.unity.inputsystem",
+            "expression": "1.1.0",
+            "define": "MLA_INPUT_SYSTEM"
+        },
+        {
+            "name": "com.unity.inputsystem",
+            "expression": "1.1.0",
+            "define": "MLA_INPUT_TESTS"
+        }
+    ],
+    "noEngineReferences": false
+}
