+++ com.unity.ml-agents/Tests/Runtime/Input/Unity.ML-Agents.Runtime.Input.Tests.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3b0ef87e2efab493280f7ea1fe55e061
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
